requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
attrs==24.2.0
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
datasets==3.0.1
dill==0.3.8
docker-pycreds==0.4.0
einops==0.8.0
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
gluonts==0.14.3
gmpy2==2.1.5
h2==4.1.0
hpack==4.0.0
huggingface_hub==0.25.2
hydra-core==1.3.2
hyperframe==6.0.1
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lightning-utilities==0.11.7
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==3.0.1
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
munkres==1.1.4
natsort==7.1.1
networkx==3.4.1
numba==0.60.0
numpy==1.26.4
omegaconf==2.3.0
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pip==24.2
platformdirs==4.3.6
propcache==0.2.0
protobuf==4.25.3
psutil==6.0.0
pyarrow==17.0.0
pycparser==2.22
pydantic==1.10.17
pygame==2.6.1
Pygments==2.18.0
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.9.0
pytorch-lightning==2.0.4
pytz==2024.1
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.9.2
safetensors==0.4.5
scikit-base==0.8.2
scikit-learn==1.5.2
scipy==1.14.1
sentry-sdk==2.16.0
setproctitle==1.3.3
setuptools==75.1.0
six==1.16.0
sktime==0.33.0
smmap==5.0.0
soundfile==0.12.1
sympy==1.13.3
threadpoolctl==3.5.0
timm==0.5.4
tokenizers==0.20.1
toolz==0.12.1
torch==2.2.0
torchaudio==2.2.0
torchmetrics==1.4.2
torchtext==0.6.0
torchvision==0.17.0
tornado==6.4.1
tqdm==4.66.5
transformers==4.45.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wandb==0.18.1
wheel==0.44.0
xxhash==3.5.0
yarl==1.14.0
zipp==3.20.2
zstandard==0.23.0
